refactor: 将 QueryContext 和 SchemaAttributePaths 从 jdbc 包迁移到 core.constructor 包
移动 QueryContext、SchemaAttributePaths 及其实现类至 constructor 包，
更新所有引用处的 import 语句
