+++ nextentity-core/src/main/java/io/github/nextentity/core/constructor/DefaultSchemaAttributePaths.java
@@ -1,4 +1,4 @@
-package io.github.nextentity.jdbc;
+package io.github.nextentity.core.constructor;
 
 import java.util.HashMap;
