+++ nextentity-core/src/main/java/io/github/nextentity/core/constructor/EmptySchemaAttributePaths.java
@@ -1,4 +1,4 @@
-package io.github.nextentity.jdbc;
+package io.github.nextentity.core.constructor;
 
 class EmptySchemaAttributePaths implements SchemaAttributePaths {
     EmptySchemaAttributePaths() {
