+++ nextentity-core/src/main/java/io/github/nextentity/core/interceptor/ConstructInterceptor.java
@@ -1,7 +1,7 @@
 package io.github.nextentity.core.interceptor;
 
 import io.github.nextentity.jdbc.Arguments;
-import io.github.nextentity.jdbc.QueryContext;
+import io.github.nextentity.core.constructor.QueryContext;
 
 /// 对象构造拦截器 - 拦截单个对象创建
 ///
