+++ nextentity-core/src/main/java/io/github/nextentity/jdbc/DefaultSqlBuilder.java
@@ -1,5 +1,6 @@
 package io.github.nextentity.jdbc;
 
+import io.github.nextentity.core.constructor.QueryContext;
 import io.github.nextentity.core.expression.ExpressionNode;
 import io.github.nextentity.core.meta.EntitySchema;
 import io.github.nextentity.core.meta.EntityType;
