+++ nextentity-core/src/test/java/io/github/nextentity/jdbc/SchemaAttributePathsTest.java
@@ -1,5 +1,6 @@
 package io.github.nextentity.jdbc;
 
+import io.github.nextentity.core.constructor.SchemaAttributePaths;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
